refactor: rename entities/auth-session to entities/auth
