+++ src/features/card/edit/model/useEditCard.ts
@@ -1,6 +1,6 @@
 import type { Card, CardEdit } from "@/entities/card";
 
-import { useAuthSession } from "@/entities/auth-session";
+import { useAuthSession } from "@/entities/auth";
 import { editCard } from "../api/editCard";
 
 type CardPatch = Omit<CardEdit, "id" | "uid">;
