+++ src/features/deck/import/hooks/useSampleDeckBootstrap.spec.tsx
@@ -21,7 +21,7 @@ const mocks = vi.hoisted(() => ({
   addSample: vi.fn<() => Promise<unknown>>(),
 }));
 
-vi.mock("@/entities/auth-session", () => ({ useAuthSession: () => mocks.auth }));
+vi.mock("@/entities/auth", () => ({ useAuthSession: () => mocks.auth }));
 vi.mock("./useDeckImport", () => ({
   useDeckImport: () => ({ addSample: mocks.addSample }),
 }));
