+++ src/pages/settings/ui/SettingsPage.spec.tsx
@@ -11,7 +11,7 @@ const mocks = vi.hoisted(() => ({
   setDarkMode: vi.fn(),
 }));
 
-vi.mock("@/entities/auth-session", () => ({ useAuthSession: () => ({ status: "initializing" as const }) }));
+vi.mock("@/entities/auth", () => ({ useAuthSession: () => ({ status: "initializing" as const }) }));
 vi.mock("@/entities/preferences", () => ({
   usePreferences: () => mocks.preferences,
   setDarkMode: mocks.setDarkMode,
